Add files via upload
Complaints registered through different ways leads to huge amount of data. To
manage and classify these huge railway data is a cubersome task, to divide them into
categories and subcategories and which complaint belong under which department.
Before getting the system to work on a complaint, a tough task is to find out which
complaints needs how much attention. Software processes the information based on
keywords to analyse the sentiments. Once analysed, the data are classified into
categories and sub-categories and than based on the particular class the complaints
get registered successfully.Therefore, through this classification the services get
simulated.